// tcru-front-sale/src/app/content/user/addaddress/interfaces/address.ts
export interface Address {
    address_id: number;
    address: string;
    user_id: number;
    amphures_id: number;
    districts_id: number;
    postal_code: string;
    province_id: number;
    geographic_id: number;
    status?: any;
    telephone: string;
    created_at?: any;
    updated_at?: any;
    firstname: string;
    lastname: string;
}